refactor src directory structure and update imports
