+++ src/app/(base)/about/content.ts
