+++ src/app/(base)/about/page.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
-import PageHeading from "@/ui/PageHeading";
-import UnderConstruction from "@/ui/UnderConstruction";
+import PageHeading from "@/components/page-heading";
+import UnderConstruction from "@/components/under-construction";
 import { aboutText, aboutImage, loremIpsum } from "./content";
 import aboutPic from "public/about-kb.jpeg";
