+++ src/app/(music)/music/@music/MusicScene.tsx
@@ -1,4 +1,4 @@
-import Canvas from "@/ui/Canvas";
+import Canvas from "@/components/canvas";
 import MusicSphere from "./MusicSphere";
 
 export default function MusicScene() {
